+++ Changelog.md
@@ -17,6 +17,7 @@
 * Add parsing of properties of table - inside horizontal, vertical borders
 * Add support of parsing more text directions in cell
 * Add support of `show_category_name` and `show_series_name` to chart label properties
+* Add method to 3 base formats to check if it contain any user data `#with_data?`
 
 ### Fixes
 * Fix parsing document style id - it can be string, not only digit